+++ rpc/rpc.go
@@ -1,0 +1,8 @@
+package rpc
+
+const (
+	packRegister = iota
+	heartBeat
+	heartBeatRet
+	packData
+)